requirements.txt
streamlit==1.42.0
groq==0.15.0
langchain==0.3.19
langchain-community==0.3.18
langchain-pinecone
langchain-huggingface==0.1.2
pinecone==5.4.2
sentence-transformers==3.4.1
pypdf==5.3.1
python-dotenv==1.0.1
